+++ src/models/Notification.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2023 Karl Pickett / ViLark Project
+// Copyright (C) 2023 Karl Pickett / Vilark Project
 namespace vilark;
 
 // A grab-bag of message types that we can send to the main thread